feature: split config responsibilities and simplify it
Config is currently a complex state store that mutates throughout the
code execution which makes it very hard to understand what is happening.
I'm splitting its responsibilities into smaller subcomponents that will
each perform a single, idempotent task whenever possible.

In this commit:
- Removes support for all package managers except npm
- Removed latitude json config from config global, now we get it from await findOrCreateConfigFile
- Made config file a module-level singleton
- Stop passing around config as prop everywhere, the point of it being a global is that you don't need to inject it everywhere

+++ packages/cli/src/lib/setupApp/index.ts
@@ -1,25 +1,27 @@
+import config from '$src/config'
+import findOrCreateConfigFile from '../latitudeConfig/findOrCreate'
+import installDependencies from './installDependencies'
+import installLatitudeServer from '../installLatitudeServer'
 import path from 'path'
-import { CLIConfig } from '$src/config'
-import cloneAppFromNpm from './cloneAppFromNpm'
-import symlinkAppFromLocal from './symlinkAppFromLocal'
-import installAppDependencies from './installDependencies'
-import updateVersion from '../latitudeConfig/updateVersion'
+import { createMasterKey } from '$src/commands/credentials/createMasterKey'
+import { existsSync, writeFileSync } from 'fs'
 import { onError } from '$src/utils'
-import { writeFileSync } from 'fs'
 
-export type Props = { version: string }
+export type Props = { version?: string }
 
 // Adds a package.json file to the app directory in development environment so
 // that you can run the cli in development mode with `npm run latitude-dev`
 function addPackageJson() {
+  if (config.pro) return
+
+  const packageJsonPath = path.resolve(config.rootDir, 'package.json')
+  if (existsSync(packageJsonPath)) return
+
   const packageJson = {
     scripts: {
       'latitude-dev': '../../node_modules/.bin/latitude',
     },
   }
-
-  const config = CLIConfig.getInstance()
-  const packageJsonPath = path.resolve(config.source, 'package.json')
 
   try {
     writeFileSync(packageJsonPath, JSON.stringify(packageJson, null, 2))
@@ -33,42 +35,10 @@ function addPackageJson() {
   }
 }
 
-export default async function setupApp({ version }: Props) {
-  const config = CLIConfig.getInstance()
-  const isPro = config.pro || config.simulatedPro
-  const setup = isPro ? cloneAppFromNpm : symlinkAppFromLocal
-
-  await setup({ version })
-
-  process.chdir(path.resolve(config.source))
-
-  if (!isPro) return addPackageJson()
-
-  try {
-    await installAppDependencies()
-  } catch (e) {
-    onError({
-      error: e as Error,
-      message: `
-      🚨 Failed to install dependencies
-
-      Update latitude and try again. If this does not solve the problem,
-      please open an issue on GitHub:
-      https://gitub.com/latitude-dev/latitude/issues
-      `,
-    })
-
-    process.exit(1)
-  }
-
-  try {
-    await updateVersion({ appDir: config.source, version })
-  } catch (e) {
-    onError({
-      error: e as Error,
-      message: '🚨 Failed to update app version',
-    })
-
-    process.exit(1)
-  }
+export default async function setupApp() {
+  const config = await findOrCreateConfigFile()
+  await installLatitudeServer({ version: config.data.version })
+  await installDependencies()
+  addPackageJson()
+  createMasterKey()
 }
